Don't sync ACLs by default during release syncing
This was mainly added for the service, so we will pass this argument
from there and in CLI, by default not sync the ACLS.
Fixes #2206
