+++ tests/integration/test_using_cockpit.py
@@ -59,7 +59,7 @@ def test_update_on_cockpit_ostree(cockpit_ostree):
     flexmock(DistGit).should_receive("existing_pr").and_return(None)
     flexmock(
         PackitAPI,
-        push_and_create_pr=lambda pr_title, pr_description, git_branch, repo: None,
+        push_and_create_pr=lambda pr_title, pr_description, git_branch, repo, sync_acls: None,
     )
 
     pc = get_local_package_config(str(upstream_path))
