+++ tests/unit/test_cli.py
@@ -61,6 +61,7 @@ def test_propose_downstream_command():
         use_downstream_specfile=False,
         package_config=PackageConfig,
         resolved_bugs=None,
+        sync_acls=False,
     ).and_return()
     result = call_packit(packit_base, parameters=["propose-downstream", "."])
     assert result.exit_code == 0
@@ -84,6 +85,7 @@ def test_pull_from_upstream_command():
         use_downstream_specfile=True,
         package_config=PackageConfig,
         resolved_bugs=None,
+        sync_acls=False,
     ).and_return()
     result = call_packit(packit_base, parameters=["pull-from-upstream", "."])
     assert result.exit_code == 0
